Bug Fixes Part 1
That's enough changes for one commit. Time to let Miris see what I've done.
