+++ hxe/kernel/src/HXE.csproj
@@ -68,9 +68,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Common\Convert.cs" />
-    <Compile Include="Common\Path.cs" />
-    <Compile Include="Common\Process.cs" />
+    <Compile Include="Common\ExConvert.cs" />
+    <Compile Include="Common\ExPath.cs" />
+    <Compile Include="Common\ExProcess.cs" />
     <Compile Include="HCE\Registry.cs" />
     <Compile Include="HCE\Progress.cs" />
     <Compile Include="NewProfile.cs" />
