+++ spv3/loader/src/Configuration.UserControl.xaml
@@ -220,7 +220,8 @@
         <DockPanel>
           <Label Content="Post-Processing:" />
           <CheckBox IsChecked="{Binding Loader.Shaders}"
-                    Unchecked="Update_AdaptiveHDRIsReady"
+                    Unchecked="Update_AdaptiveHDR_isReady"
+                    Checked="Update_AdaptiveHDR_isReady"
                     ToolTip="Enables SPV3's post-processing system." />
         </DockPanel>
         <DockPanel>
@@ -229,6 +230,7 @@
           <CheckBox IsChecked="{Binding Shaders.AdaptiveHDR}"
                     IsEnabled="{Binding Loader.Shaders}"
                     Unchecked="AdaptiveHDR_Unchecked"
+                    Checked="Update_AdaptiveHDR_isReady"
                     ToolTip="Enables High Dynamic Range. Adapts to the scene's light levels. Required for Bloom and Screen Space Reflections to render corectly." />
         </DockPanel>
         <DockPanel>
@@ -304,7 +306,7 @@
           <Label Content="SSR:"
                  IsEnabled="{Binding Loader.Shaders}" />
           <CheckBox IsChecked="{Binding Shaders.SSR}"
-                    IsEnabled="{Binding AdaptiveHDRIsReady}"
+                    IsEnabled="{Binding Shaders.AdaptiveHDR_isReady}"
                     Checked="SSR_Or_ResolutionChanged"
                     ToolTip="Requires Adaptive HDR. Enables Screen Space Ray Traced Reflections." />
         </DockPanel>
@@ -400,7 +402,7 @@
           <Label Content="G-Buffer:" />
           <CheckBox IsChecked="{Binding OpenSauce.GBuffer}"
                     ToolTip="Disabling this will deactivate VISR and Thermal Vision."
-                    Unchecked="GBufferUnchecked" />
+                    Unchecked="GBuffer_Unchecked" />
         </DockPanel>
         <DockPanel>
           <Label Content="Normal Maps:" />
@@ -425,7 +427,7 @@
         <DockPanel>
           <Label Content="Bloom:" />
           <CheckBox IsChecked="{Binding OpenSauce.Bloom}"
-                    IsEnabled="{Binding AdaptiveHDRIsReady}"
+                    IsEnabled="{Binding Shaders.AdaptiveHDR_isReady}"
                     ToolTip="Requires Adaptive HDR. Adds glowy highlights to visual bright spots." />
         </DockPanel>
