+++ spv3/loader/src/Configuration.cs
@@ -34,8 +34,6 @@ namespace SPV3
     public HXE.Settings           Settings  { get; set; } = new HXE.Settings(hxe);
     public HXE.Positions          Positions { get; set; } = new HXE.Positions();
 
-    public bool AdaptiveHDRIsReady = false; // see Configuration.UserControl.xaml.cs, Update_AdaptiveHDRIsReady().
-
     public void Load()
     {
       spv3 = Loader;
